+++ app/src/main/java/com/mucsc3550/cassie/framework/HelpScreen3.java
@@ -10,7 +10,7 @@ public class HelpScreen3 extends Screen {
     }
 
     @Override
-    public void update(float deltaTime) {
+    public void update(double deltaTime) {
         List<TouchEvent> touchEvents = game.getInput().getTouchEvents();
         int len = touchEvents.size();
         for(int i = 0; i < len; i++) {
@@ -26,7 +26,7 @@ public class HelpScreen3 extends Screen {
     }
 
     @Override
-    public void present(float deltaTime) {
+    public void present(double deltaTime) {
         Graphics g = game.getGraphics();
 
         g.drawPixmap(Assets.background, 0, 0);
