+++ app/src/main/java/com/mucsc3550/cassie/framework/MainMenuScreen.java
@@ -8,7 +8,7 @@ public class MainMenuScreen extends Screen {
         super(game);
     }
 
-    public void update(float deltaTime) {
+    public void update(double deltaTime) {
         Graphics g = game.getGraphics();
         List<TouchEvent> touchEvents = game.getInput().getTouchEvents();
 
@@ -25,6 +25,7 @@ public class MainMenuScreen extends Screen {
                     if(Settings.soundEnabled) Assets.click.play(1);
                 }
                 if(inBounds(event, 64, 220 + 65, 192, 55)) {
+                    //Settings.addScore(54);
                     game.setScreen(new HighScoreScreen(game));
                     if(Settings.soundEnabled) Assets.click.play(1);
                 }
@@ -37,7 +38,7 @@ public class MainMenuScreen extends Screen {
     }
 
     @Override
-    public void present(float deltaTime) {
+    public void present(double deltaTime) {
         Graphics g = game.getGraphics();
         g.drawPixmap(Assets.background, 0, 0);
         g.drawPixmap(Assets.logo, 45, 120);
