+++ app/src/main/java/com/mucsc3550/cassie/framework/impl/AndroidFastRenderView.java
@@ -28,7 +28,7 @@ public class AndroidFastRenderView extends SurfaceView implements Runnable {
         while(running) {
             if(!holder.getSurface().isValid()) continue;
 
-            float deltaTime = (System.nanoTime()-startTime) / 1000000000.0f;
+            double deltaTime = (System.nanoTime()-startTime) / 1000000000.0f;
             startTime = System.nanoTime();
 
             game.getCurrentScreen().update(deltaTime);
